CUT encrypt.py & test_conduit_encrypt.py because functionality has been ported to `Conduit.encrypt()`  - refactored `Algorithms` enum to pdfconduit.py
