+++ pdfconduit/__init__.py
@@ -1,10 +1,8 @@
-from pdfconduit.conduit import Encrypt
 from pdfconduit.transform import Merge, Rotate, Upscale, slicer
 from pdfconduit.utils import Info
 from pdfconduit.pdfconduit import Conduit
 
 __all__ = [
-    "Encrypt",
     "Merge",
     "Rotate",
     "Upscale",
