--- pdfconduit/conduit/__init__.py
@@ -1,3 +1,0 @@
-from pdfconduit.conduit.encrypt import Encrypt
-
-__all__ = ["Encrypt"]
